Remove RequestManager Class
