+++ src/Eth.php
@@ -13,10 +13,7 @@ namespace Web3;
 
 use Web3\Providers\Provider;
 use Web3\Providers\HttpProvider;
-use Web3\RequestManagers\RequestManager;
-use Web3\RequestManagers\HttpRequestManager;
 use Web3\Providers\WsProvider;
-use Web3\RequestManagers\WsRequestManager;
 
 class Eth
 {
@@ -54,13 +51,9 @@ class Eth
         if (is_string($provider) && (filter_var($provider, FILTER_VALIDATE_URL) !== false)) {
             // check the uri schema
             if (preg_match('/^https?:\/\//', $provider) === 1) {
-                $requestManager = new HttpRequestManager($provider);
-
-                $this->provider = new HttpProvider($requestManager);
+                $this->provider = new HttpProvider($provider);
             } else if (preg_match('/^wss?:\/\//', $provider) === 1) {
-                $requestManager = new WsRequestManager($provider);
-
-                $this->provider = new WsProvider($requestManager);
+                $this->provider = new WsProvider($provider);
             }
         } else if ($provider instanceof Provider) {
             $this->provider = $provider;
