+++ src/Providers/WsProvider.php
@@ -11,9 +11,9 @@
 
 namespace Web3\Providers;
 
+use React\Async;
 use Web3\Providers\Provider;
 use Web3\Providers\IProvider;
-use Web3\RequestManagers\RequestManager;
 
 class WsProvider extends Provider implements IProvider
 {
@@ -25,14 +25,35 @@ class WsProvider extends Provider implements IProvider
     protected $methods = [];
 
     /**
+     * client
+     *
+     * @var \Web3\Providers\WsClient
+     */
+    protected $client;
+
+    /**
      * construct
-     * 
-     * @param \Web3\RequestManagers\RequestManager $requestManager
+     *
+     * @param string $host
+     * @param int $timeout
      * @return void
      */
-    public function __construct(RequestManager $requestManager)
+    public function __construct($host, $timeout = 1)
     {
-        parent::__construct($requestManager);
+        parent::__construct($host, $timeout);
+        $this->client = new WsClient(
+            $host,
+            function ($obj, $message) {
+            },
+            function ($obj, $error) {
+            },
+            function ($obj, $close) {
+            },
+            function ($obj, $connected) {
+            },
+            $timeout
+        );
+        $this->client->set_ws_connector();
     }
 
     /**
@@ -42,7 +63,7 @@ class WsProvider extends Provider implements IProvider
      */
     public function close()
     {
-        $this->requestManager->close();
+        $this->client->close();
     }
 
     /**
@@ -69,7 +90,7 @@ class WsProvider extends Provider implements IProvider
 
                 return call_user_func($callback, null, $res);
             };
-            return $this->requestManager->sendPayload($payload, $proxy);
+            return $this->sendPayload($payload, $proxy);
         } else {
             $this->methods[] = $method;
             $this->batch[] = $payload;
@@ -118,9 +139,80 @@ class WsProvider extends Provider implements IProvider
             }
             return call_user_func($callback, null, $res);
         };
-        $r = $this->requestManager->sendPayload('[' . implode(',', $this->batch) . ']', $proxy);
+        $r = $this->sendPayload('[' . implode(',', $this->batch) . ']', $proxy);
         $this->methods = [];
         $this->batch = [];
         return $r;
+    }
+
+    /**
+     * sendPayload
+     *
+     * @param string $payload
+     * @param callable $callback
+     * @return void
+     */
+    public function sendPayload($payload, $callback)
+    {
+        if (!is_string($payload)) {
+            throw new \InvalidArgumentException('Payload must be string.');
+        }
+
+        if (!$this->client->isConnected) {
+            Async\await($this->client->connect(0));
+        }
+
+        $host = $this->host;
+        $request = function () use ($host, $payload, $callback) {
+            try {
+                $res = Async\await($this->client->send($payload));
+                $json = json_decode($res);
+
+                if (JSON_ERROR_NONE !== json_last_error()) {
+                    call_user_func($callback, new InvalidArgumentException('json_decode error: ' . json_last_error_msg()), null);
+                }
+                if (is_array($json)) {
+                    // batch results
+                    $results = [];
+                    $errors = [];
+
+                    foreach ($json as $result) {
+                        if (property_exists($result,'result')) {
+                            $results[] = $result->result;
+                        } else {
+                            if (isset($json->error)) {
+                                $error = $json->error;
+                                $errors[] = new RPCException(mb_ereg_replace('Error: ', '', $error->message), $error->code);
+                            } else {
+                                $results[] = null;
+                            }
+                        }
+                    }
+                    if (count($errors) > 0) {
+                        call_user_func($callback, $errors, $results);
+                    } else {
+                        call_user_func($callback, null, $results);
+                    }
+                } elseif (property_exists($json,'result')) {
+                    call_user_func($callback, null, $json->result);
+                } else {
+                    if (isset($json->error)) {
+                        $error = $json->error;
+
+                        call_user_func($callback, new RPCException(mb_ereg_replace('Error: ', '', $error->message), $error->code), null);
+                    } else {
+                        call_user_func($callback, new RPCException('Something wrong happened.'), null);
+                    }
+                }
+            } catch (Exception $err) {
+                call_user_func($callback, $err, null);
+            }
+        };
+
+        if (function_exists('React\\Async\\async')) {
+            $request = Async\async($request);
+        }
+
+        return Async\coroutine($request);
     }
 }
