+++ test/TestCase.php
@@ -4,7 +4,7 @@ namespace Test;
 
 use \PHPUnit\Framework\TestCase as BaseTestCase;
 use Web3\Web3;
-use Web3\RequestManagers\HttpAsyncRequestManager;
+use Web3\Providers\HttpAsyncProvider;
 use Web3\Providers\HttpProvider;
 
 class TestCase extends BaseTestCase
@@ -59,8 +59,7 @@ class TestCase extends BaseTestCase
         $web3 = new Web3($this->testHost);
         $this->web3 = $web3;
 
-        $asyncRequestManager = new HttpAsyncRequestManager($this->testHost);
-        $asyncHttpProvider = new HttpProvider($asyncRequestManager);
+        $asyncHttpProvider = new HttpAsyncProvider($this->testHost);
         $this->asyncHttpProvider = $asyncHttpProvider;
 
         $web3->eth->coinbase(function ($err, $coinbase) {
