+++ test/unit/HttpProviderTest.php
@@ -4,8 +4,7 @@ namespace Test\Unit;
 
 use RuntimeException;
 use Test\TestCase;
-use Web3\RequestManagers\HttpRequestManager;
-use Web3\RequestManagers\HttpAsyncRequestManager;
+use Web3\Providers\HttpAsyncProvider;
 use Web3\Providers\HttpProvider;
 use Web3\Methods\Web3\ClientVersion;
 
@@ -18,8 +17,7 @@ class HttpProviderTest extends TestCase
      */
     public function testSend()
     {
-        $requestManager = new HttpRequestManager($this->testHost);
-        $provider = new HttpProvider($requestManager);
+        $provider = new HttpProvider($this->testHost);
         $method = new ClientVersion('web3_clientVersion', []);
 
         $provider->send($method, function ($err, $version) {
@@ -37,8 +35,7 @@ class HttpProviderTest extends TestCase
      */
     public function testBatch()
     {
-        $requestManager = new HttpRequestManager($this->testHost);
-        $provider = new HttpProvider($requestManager);
+        $provider = new HttpProvider($this->testHost);
         $method = new ClientVersion('web3_clientVersion', []);
         $callback = function ($err, $data) {
             if ($err !== null) {
@@ -66,8 +63,7 @@ class HttpProviderTest extends TestCase
      */
     public function testSendAsync()
     {
-        $requestManager = new HttpAsyncRequestManager($this->testHost);
-        $provider = new HttpProvider($requestManager);
+        $provider = new HttpAsyncProvider($this->testHost);
         $method = new ClientVersion('web3_clientVersion', []);
 
         // \React\Async\await($provider->send($method, function ($err, $version) {
@@ -108,8 +104,7 @@ class HttpProviderTest extends TestCase
      */
     public function testBatchAsync()
     {
-        $requestManager = new HttpAsyncRequestManager($this->testHost);
-        $provider = new HttpProvider($requestManager);
+        $provider = new HttpAsyncProvider($this->testHost);
         $method = new ClientVersion('web3_clientVersion', []);
         $callback = function ($err, $data) {
             if ($err !== null) {
